refactor: UserInfoModal 내의 도메인 로직을 커스텀 훅으로 분리
(cherry picked from commit 75ff601c50abfcd49d879c3543f6d3ec08cf3749)
